scraper/requirements.txt
playwright==1.44.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
